adding robust ID support to plotting utils
